+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,84 @@
+# @forestadmin/agent [1.28.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.27.1...@forestadmin/agent@1.28.0) (2023-09-12)
+
+
+### Features
+
+* **widgets:** add support for currency widget ([#812](https://github.com/ForestAdmin/agent-nodejs/issues/812)) ([2f263bb](https://github.com/ForestAdmin/agent-nodejs/commit/2f263bb2d263a6e942a349ec7744284837e01ad4))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.28.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.21.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.17.0
+
+## @forestadmin/agent [1.27.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.27.0...@forestadmin/agent@1.27.1) (2023-09-08)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.27.1
+* **@forestadmin/datasource-toolkit:** upgraded to 1.20.1
+* **@forestadmin/forestadmin-client:** upgraded to 1.16.1
+
+# @forestadmin/agent [1.27.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.26.0...@forestadmin/agent@1.27.0) (2023-09-07)
+
+
+### Features
+
+* **widgets:** add support for color picker in actions ([bae889c](https://github.com/ForestAdmin/agent-nodejs/commit/bae889c144e6d24d58a601f1ea4dea658889b321))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.27.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.20.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.16.0
+
+# @forestadmin/agent [1.26.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.25.0...@forestadmin/agent@1.26.0) (2023-09-05)
+
+
+### Features
+
+* **widgets:** add support for number input lists with options in actions ([519d1b7](https://github.com/ForestAdmin/agent-nodejs/commit/519d1b7c344456cdcf26a8972cef619c4efb1a48))
+* **widgets:** add support for number input lists with options in actions ([#807](https://github.com/ForestAdmin/agent-nodejs/issues/807)) ([22d84bd](https://github.com/ForestAdmin/agent-nodejs/commit/22d84bd104dcc7893a9536365329db5ee2467c28))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.26.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.19.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.15.0
+
+# @forestadmin/agent [1.25.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.24.0...@forestadmin/agent@1.25.0) (2023-09-05)
+
+
+### Features
+
+* **widgets:** add support for number input in custom actions ([8f52fb3](https://github.com/ForestAdmin/agent-nodejs/commit/8f52fb3c90e5050873390a167fbb11df5fa34863))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.25.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.18.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.14.0
+
 # @forestadmin/agent [1.24.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.23.0...@forestadmin/agent@1.24.0) (2023-09-01)