+++ packages/plugin-flattener/CHANGELOG.md
@@ -1,3 +1,58 @@
+## @forestadmin/plugin-flattener [1.0.48](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.47...@forestadmin/plugin-flattener@1.0.48) (2023-09-12)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.21.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.28.0
+
+## @forestadmin/plugin-flattener [1.0.47](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.46...@forestadmin/plugin-flattener@1.0.47) (2023-09-08)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.20.1
+* **@forestadmin/datasource-customizer:** upgraded to 1.27.1
+
+## @forestadmin/plugin-flattener [1.0.46](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.45...@forestadmin/plugin-flattener@1.0.46) (2023-09-07)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.20.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.27.0
+
+## @forestadmin/plugin-flattener [1.0.45](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.44...@forestadmin/plugin-flattener@1.0.45) (2023-09-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.19.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.26.0
+
+## @forestadmin/plugin-flattener [1.0.44](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.43...@forestadmin/plugin-flattener@1.0.44) (2023-09-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.18.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.25.0
+
 ## @forestadmin/plugin-flattener [1.0.43](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.42...@forestadmin/plugin-flattener@1.0.43) (2023-09-01)